+++ src/control_plane/migration/mod.rs
@@ -8,6 +8,7 @@ mod migration_execution_result;
 mod migration_operation_error;
 mod migration_operations;
 mod migration_rollback_plan;
+mod migration_target_plan;
 mod postgres;
 mod run_migration;
 
@@ -18,4 +19,5 @@ pub(crate) use migration_execution_result::MigrationExecutionResult;
 pub(crate) use migration_operation_error::MigrationOperationError;
 pub(crate) use migration_operations::{MigrationFuture, MigrationOperations};
 pub(crate) use migration_rollback_plan::MigrationRollbackPlan;
+pub(crate) use migration_target_plan::MigrationTargetPlan;
 pub(crate) use run_migration::{confirm_migration, execute_migration, rollback_migration};